# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sincronismo de dados entre o Controle de Acesso e o Refeitório.",
        "title": "API de Integração - Módulo Refeitório",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/integracao/funcionario": {
            "post": {
                "description": "Recebe os dados cadastrais de um funcionário vindo do controle de acesso e grava no refeitório",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Integração"
                ],
                "summary": "Integrar Funcionário",
                "parameters": [
                    {
                        "description": "Dados do Funcionário",
                        "name": "funcionario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Funcionario"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    }
                }
            }
        },
        "/api/integracao/terceiro": {
            "post": {
                "description": "Recebe os dados de um terceiro vindo do controle de acesso e grava no refeitório",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Integração"
                ],
                "summary": "Integrar Terceiro",
                "parameters": [
                    {
                        "description": "Dados do Terceiro",
                        "name": "terceiro",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Terceiro"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    }
                }
            }
        },
        "/api/integracao/visitante": {
            "post": {
                "description": "Recebe os dados de um visitante vindo do controle de acesso e grava no refeitório",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Integração"
                ],
                "summary": "Integrar Visitante",
                "parameters": [
                    {
                        "description": "Dados do Visitante",
                        "name": "visitante",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Visitante"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespostaPadrao"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.RespostaPadrao": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "mensagem": {
                    "type": "string"
                },
                "sucesso": {
                    "type": "boolean"
                }
            }
        },
        "models.Empresa": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "models.Equipamento": {
            "type": "object",
            "properties": {
                "empresa": {
                    "$ref": "#/definitions/models.Empresa"
                },
                "empresaID": {
                    "type": "integer"
                },
                "idequip": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "models.Funcionario": {
            "type": "object",
            "properties": {
                "admissao": {
                    "type": "string"
                },
                "ativo": {
                    "type": "string"
                },
                "cargo": {
                    "type": "string"
                },
                "centroDeCusto": {
                    "type": "string"
                },
                "credencial": {
                    "type": "string"
                },
                "departamento": {
                    "type": "string"
                },
                "documento": {
                    "type": "string"
                },
                "empresa": {
                    "$ref": "#/definitions/models.Empresa"
                },
                "empresaID": {
                    "type": "integer"
                },
                "equipamentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipamento"
                    }
                },
                "grupoRefeicao": {
                    "$ref": "#/definitions/models.GrupoRefeicao"
                },
                "grupoRefeicaoID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "tipoPessoa": {
                    "type": "integer"
                }
            }
        },
        "models.GrupoRefeicao": {
            "type": "object",
            "properties": {
                "empresa": {
                    "$ref": "#/definitions/models.Empresa"
                },
                "empresaID": {
                    "type": "integer"
                },
                "idgrup": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "refeicoes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Refeicao"
                    }
                }
            }
        },
        "models.PrecoRefeicao": {
            "type": "object",
            "properties": {
                "dataFim": {
                    "description": "NULL = Valor ativo atual",
                    "type": "string"
                },
                "dataInicio": {
                    "type": "string"
                },
                "idpreco": {
                    "type": "integer"
                },
                "refeicaoID": {
                    "type": "integer"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "models.Refeicao": {
            "type": "object",
            "properties": {
                "empresa": {
                    "$ref": "#/definitions/models.Empresa"
                },
                "empresaID": {
                    "type": "integer"
                },
                "historicoPrecos": {
                    "description": "Relacionamento com o historico de valores comerciais",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PrecoRefeicao"
                    }
                },
                "horaFim": {
                    "type": "string"
                },
                "horaInicio": {
                    "type": "string"
                },
                "idref": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "models.Terceiro": {
            "type": "object",
            "properties": {
                "credencial": {
                    "type": "string"
                },
                "dataFim": {
                    "type": "string"
                },
                "dataInicio": {
                    "type": "string"
                },
                "documento": {
                    "type": "string"
                },
                "empTer": {
                    "type": "string"
                },
                "empresa": {
                    "$ref": "#/definitions/models.Empresa"
                },
                "empresaID": {
                    "type": "integer"
                },
                "equipamentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipamento"
                    }
                },
                "funcionario": {
                    "$ref": "#/definitions/models.Funcionario"
                },
                "funcionarioID": {
                    "type": "integer"
                },
                "grupoRefeicao": {
                    "$ref": "#/definitions/models.GrupoRefeicao"
                },
                "grupoRefeicaoID": {
                    "type": "integer"
                },
                "horaFim": {
                    "type": "string"
                },
                "horaInicio": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "tipoPessoa": {
                    "type": "integer"
                }
            }
        },
        "models.Visitante": {
            "type": "object",
            "properties": {
                "credencial": {
                    "type": "string"
                },
                "dataFim": {
                    "type": "string"
                },
                "dataInicio": {
                    "type": "string"
                },
                "documento": {
                    "type": "string"
                },
                "empresa": {
                    "$ref": "#/definitions/models.Empresa"
                },
                "empresaID": {
                    "type": "integer"
                },
                "equipamentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipamento"
                    }
                },
                "funcionario": {
                    "$ref": "#/definitions/models.Funcionario"
                },
                "funcionarioID": {
                    "type": "integer"
                },
                "grupoRefeicao": {
                    "$ref": "#/definitions/models.GrupoRefeicao"
                },
                "grupoRefeicaoID": {
                    "type": "integer"
                },
                "horaFim": {
                    "type": "string"
                },
                "horaInicio": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "motivo": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "tipoPessoa": {
                    "type": "integer"
                }
            }
        }
    }
}